Updated form-data service. Updating hardware implemented
Form data service was updated. Should only contain functions that are specific to that servce. Updated 'hardware-uitlenen-form'. Should be complete. Now updates database with status and adds a new document in lendings. Code is also more optimized. Also added a reset button to reset all hardware status to available for testing. Can be turned off by changing a boolean in the service